+++ src/routes/account-export-recovery.routes.js
@@ -1,6 +1,8 @@
+// @ts-check
+
 import { Router } from "express";
 import { accountExportRecoveryService } from "../services/account-export-recovery.service.js";
-import { asyncRoute } from "../utils/http.js";
+import { authenticatedAsyncRoute as asyncRoute } from "../utils/http.js";
 
 const accountExportRecoveryRoutes = Router();