+++ src/routes/jobs.routes.js
@@ -1,6 +1,8 @@
+// @ts-check
+
 import { Router } from "express";
 import { jobsService } from "../services/jobs.service.js";
-import { asyncRoute } from "../utils/http.js";
+import { workspaceAsyncRoute as asyncRoute } from "../utils/http.js";
 
 const jobsRoutes = Router();